feat[#232] : 도메인 리팩토링에 따른 충돌 해결
