+++ src/features/marking/api/addMarking.ts
@@ -1,10 +1,12 @@
 import { useMutation } from "@tanstack/react-query";
 import { LatLng } from "@/features/auth/api/region";
+import { useMapStore } from "@/features/map/store";
 import {
   MARKING_ADD_ERROR_MESSAGE,
   MARKING_REQUEST_URL,
   POST_VISIBILITY_MAP,
 } from "../constants";
+import { useMarkingFormStore } from "../store";
 
 // Marking Form 저장 API
 export interface MarkingFormRequest extends LatLng {
@@ -59,8 +61,22 @@ const postMarkingFormData = async ({
   return data;
 };
 
-export const usePostMarkingForm = () => {
+export const usePostMarkingForm = ({
+  onSuccess,
+}: {
+  onSuccess: () => void;
+}) => {
+  const resetMarkingFormStore = useMarkingFormStore(
+    (state) => state.resetMarkingFormStore,
+  );
+  const setMode = useMapStore((state) => state.setMode);
+
   return useMutation({
     mutationFn: postMarkingFormData,
+    onSuccess: () => {
+      onSuccess();
+      resetMarkingFormStore();
+      setMode("view");
+    },
   });
 };
