+++ src/features/marking/ui/markingFormModal.tsx
@@ -320,7 +320,12 @@ const SaveButton = ({ onCloseMarkingModal }: MarkingFormModalProps) => {
     ),
   );
 
-  const { mutate: postMarkingData } = usePostMarkingForm();
+  const { mutate: postMarkingData } = usePostMarkingForm({
+    onSuccess: () => {
+      onCloseMarkingModal();
+      onOpenSnackbar();
+    },
+  });
 
   const handleSave = () => {
     const { token } = useAuthStore.getState();
@@ -378,11 +383,6 @@ const TemporarySaveButton = ({
 }: MarkingFormModalProps) => {
   const map = useMap();
 
-  const setMode = useMapStore((state) => state.setMode);
-  const resetMarkingFormStore = useMarkingFormStore(
-    (state) => state.resetMarkingFormStore,
-  );
-
   const { handleOpen: onOpenSnackbar, onClose: onCloseSnackbar } = useSnackBar(
     () => (
       <MapSnackbar onClose={onCloseSnackbar}>
@@ -395,8 +395,6 @@ const TemporarySaveButton = ({
   const { mutate: postTempMarkingData } = usePostTempMarkingForm({
     onSuccess: () => {
       onCloseMarkingModal();
-      resetMarkingFormStore();
-      setMode("view");
       onOpenSnackbar();
     },
   });
